Reaction variation changed to reaction model
